[EMR][Feature] Support get ip by host for stream load in some enviroment like k8s
In order to adapt to the k8s environment, our BE registration logic currently uses hostname directly. However, this hostname cannot be recognized in the external environment, so an unknownhost exception like 'java.net.UnknownHostException: backend-1' will be thrown when importing data using stream load. This pr supports obtaining ip according to hostname and returning it to the client
        Link: https://code.alibaba-inc.com/soe/starrocks/codereview/11447359

    * [Feature] Support get ip by host for stream load in some enviroment like k8s